+++ packages/core/src/Components/InputText/Types/InputText.ts
@@ -1,4 +1,4 @@
-import type { LabelType }   from '@/Components/Label';
+import type { LabelType }   from '@/Components/Label/Internal';
 import type { Shape, Size } from '@/Shared/Types';
 
 export interface InputTextLabelProp {
