+++ packages/core/src/Components/Link/UI/FoLink.vue
@@ -9,7 +9,7 @@
 
 <script setup lang="ts">
 import type { LinkProps, UnderlineLinkEffect } from '@/Components/Link/Types/Link';
-import FoRouterLink                            from '@/Components/Link/Internal';
+import { FoRouterLink }                        from '@/Components/Link/Internal';
 import { useColor }                            from '@/Shared/Internal/Lib';
 import { computed }                            from 'vue';
